moved every folder to src/movie_booking_fastapi
