+++ pyproject.toml
@@ -31,4 +31,4 @@ allow-direct-references = true
 packages = ["src/movie_booking_fastapi"]
 
 [tool.rye.scripts]
-start = "fastapi dev src/movie_booking_fastapi/server.py"
+start = "fastapi dev src/movie_booking_fastapi/main.py"
