+++ src/movie_booking_fastapi/database/__init__.py
@@ -1,0 +1,1 @@
+from .models import *
