+++ src/movie_booking_fastapi/main.py
@@ -4,19 +4,22 @@ from fastapi.responses import JSONResponse
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from typing import Optional,Annotated,List
 from datetime import date, datetime,timedelta,timezone
-from database.database import engine, SessionLocal, get_db
+from .database.database import engine, SessionLocal, get_db
 from jwt.exceptions import InvalidTokenError
 from sqlalchemy.orm import Session
-from handlers import movie_handler, user_handler,reserve_handler
-from model import *
-from database import models
-from database.models import Users
-from logger import *
+from .handlers import movie_handler, user_handler,reserve_handler
+from .auth.auth import authenticate_user
+from .model import *
+from .database import models
+from .database.models import Users
+from movie_booking_fastapi.logger import *
 from decouple import config
 from contextlib import asynccontextmanager
 import logging
-from logger import CustomLog,logger
-from exceptions import UserCreationError
+from movie_booking_fastapi.logger import CustomLog,logger
+from movie_booking_fastapi.exceptions import UserCreationError,InvalidUserNamePasswordError
+from .auth.auth import get_current_user
+
 
 
 
@@ -44,50 +47,28 @@ models.Base.metadata.create_all(bind=engine)
 @app.post('/users/',tags=["users"])
 async def create_users(user: UsersBase, db:db_dependency):
     logger.info("users endpoint accessed")
+
     user = user_handler.create_user(db,user,logger)
     if not user:
         raise UserCreationError
+    
     return user
 
-    # if user_handler.check_user_exist(db,user.username):
-    #     return JSONResponse(status_code=409, content={'detail':'Username already exist'})
-    # if user_handler.add_user(db,user):
-    #     logger.info(f"{user.username} created")
-    #     return UserResponse(**user.model_dump())
     
 
 @app.post('/login/',tags=["users"]) 
 def login_user( db:db_dependency,form_data: OAuth2PasswordRequestForm=Depends()):
     logger.info("login endpoint accessed")
-    user = user_handler.authenticate_user(db,form_data.username,form_data.password)
-    if not user:
-        return JSONResponse(status_code=401,content={'detail':'Incorrect username or password'})
-    access_token = user_handler.create_access_token(
-        data={'sub':user.username},
-        ALGORITHM=ALGORITHM,
-        SECRET_KEY=SECRET_KEY,
-        expires_delta=timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-    )
-    logger.info(f"{user.username} logged in")
-    return Token(access_token=access_token,token_type='bearer')
-    
 
-def get_current_user(db:db_dependency, token: Annotated[str, Depends(oauth2_scheme)]):
-    try:
-        payload = jwt.decode(token, SECRET_KEY,algorithms=[ALGORITHM])
-        username :str = payload.get('sub')
-        
-        if username is None:
-            return JSONResponse(status_code=404,content={'detail':"username not found"})
-        token_data = TokenData(username=username)
-        
-        # return token_data.username
-    except InvalidTokenError :
-        return JSONResponse(content={'detail':'Invalid key'})
-    user = user_handler.get_user(db,token_data.username)
-    if not user:
-        return JSONResponse(status_code=401,content={'detail':'Could not validate Credentials'})
-    return user
+    token = authenticate_user(db,form_data.username,form_data.password)
+    if not token:
+        raise InvalidUserNamePasswordError
+    
+    logger.info(f"{form_data.username} logged in")
+    return token
+
+
+
 
 @app.post('/movies/',tags=["movies"])
 async def add_movies(db:db_dependency,movie :MovieBase, current_user:Annotated[Users, Depends(get_current_user)]):
