+++ src/movie_booking_fastapi/model/movie.py
